+++ DB/tables/produtos.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS produtos (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     vendedor_id VARCHAR(20) NOT NULL,
     nome VARCHAR(255) NOT NULL,
     descricao TEXT NOT NULL,
     preco DECIMAL(10,2) NOT NULL CHECK (preco >= 0),
     quantidade INT NOT NULL CHECK (quantidade >= 0),
-    categoria_id BIGINT,
+    categoria_id INT,
     data_cadastro TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     `status` ENUM('ativo', 'inativo') DEFAULT 'ativo',
     media_avaliacoes DECIMAL(3,2) DEFAULT 0,
