+++ DB/tables/tag.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS tags (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     nome VARCHAR(50) UNIQUE NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS produto_tags (
-    produto_id BIGINT NOT NULL,
-    tag_id BIGINT NOT NULL,
+    produto_id INT NOT NULL,
+    tag_id INT NOT NULL,
     PRIMARY KEY (produto_id, tag_id),
     FOREIGN KEY (produto_id) REFERENCES produtos(id) ON DELETE CASCADE,
     FOREIGN KEY (tag_id) REFERENCES tags(id) ON DELETE CASCADE
